+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2 (unreleased)
+
+* Add support for delayed response
+* Add `Via` header
+* Update `kong.core.handler` to `kong.runloop.handler`
+
 ## Unreleased
 
 * Initial commit